optimize the code, create custom hook
