+++ src/components/pages/dashboard-pages/EditClass.jsx
@@ -1,47 +1,46 @@
-import {
-  getClassDetails,
-  getUserData,
-  updateClassForm,
-} from '../../../util/DataBaseRequests';
-import { useEffect, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
-
 import ClassForm from './ClassForm';
 import Loader from '../../common/Loader';
+import { getClassDetails } from '../../../util/DataBaseRequests';
 import { useAuth } from '../../../AuthProvider';
+import useClasses from '../../../util/ClassService';
+import { useEffect } from 'react';
+import { useParams } from 'react-router-dom';
 
 const EditClass = () => {
   const { classId } = useParams();
-  const { userData, setUserData } = useAuth();
-  const [category, setCategory] = useState(null);
-  const [formData, setFormData] = useState({
-    classTitle: '',
-    category: category,
-    description: '',
-    classImage: {},
-    price: '',
-    duration: '',
-    ages: { minAge: '', maxAge: '' },
-    type: '',
-    lessonType: '',
-    goal: '',
-    experience: '',
-    other: '',
-    availableTime: [{ date: '', startTime: '' }],
-  });
-  const [formErrors, setFormErrors] = useState({});
-  const [isLoading, setIsLoading] = useState(false);
-  const navigate = useNavigate();
+  const { userData } = useAuth();
+  const {
+    handleSubmit,
+    handleSubjects,
+    handleChange,
+    handleAddTime,
+    returnToClasses,
+    formErrors,
+    formData,
+    isLoading,
+    category,
+    setFormData,
+    setIsLoading,
+    setCategory,
+    setFormErrors,
+  } = useClasses();
 
   useEffect(() => {
     setIsLoading(true);
     const getClassData = async () => {
       try {
         const response = await getClassDetails(classId, userData.token);
-        console.log(response);
         if (response.class) {
           setFormData({
-            ...response.class,
+            classTitle: response.class.classTitle,
+            description: response.class.description,
+            price: response.class.price,
+            duration: response.class.duration,
+            type: response.class.type,
+            lessonType: response.class.lessonType,
+            goal: response.class.goal,
+            experience: response.class.experience,
+            other: response.class.other,
             ages: {
               minAge: response.class.ages.minAge,
               maxAge: response.class.ages.maxAge,
@@ -67,163 +66,6 @@ const EditClass = () => {
     getClassData();
   }, [classId, userData.token]);
 
-  console.log('formdata', formData);
-  const returnToClasses = () => navigate('/dashboard/classes');
-
-  const handleAddTime = () => {
-    setFormData((formData) => ({
-      ...formData,
-      availableTime: [...formData.availableTime, { date: '', startTime: '' }],
-    }));
-  };
-
-  const handleChange = (e, index) => {
-    const { name, value, files } = e.target;
-  
-    if (name === 'classImage') {
-      setFormData({ ...formData, classImage: files[0] });
-    } else if (name === 'minAge' || name === 'maxAge') {
-      setFormData({
-        ...formData,
-        ages: {
-          ...formData.ages,
-          [name]: value,
-        },
-      });
-    } else if (name === 'date' || name === 'startTime') {
-      const newAvailableTime = [...formData.availableTime];
-      newAvailableTime[index][name] = value;
-      setFormData({
-        ...formData,
-        availableTime: newAvailableTime,
-      });
-  
-      let availableTimeError = false;
-      newAvailableTime.forEach((time) => {
-        if (!time.date || !time.startTime) {
-          availableTimeError = true;
-        }
-      });
-  
-      setFormErrors((formErrors) => ({
-        ...formErrors,
-        availableTime: availableTimeError ? 'Please provide the date and start time for all available times' : '',
-      }));
-    } else {
-      setFormData({ ...formData, [name]: value });
-    }
-    setFormErrors((formErrors) => ({
-      ...formErrors,
-      [name]: '',
-    }));
-  };
-
-  const handleSubjects = (value) => {
-    setCategory(value);
-    setFormData({ ...formData, category: value.value });
-    setFormErrors((formErrors) => ({
-      ...formErrors,
-      category: '',
-    }));
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    let availableTimeError = false;
-    formData.availableTime.forEach((time) => {
-      if (!time.date || !time.startTime) {
-        availableTimeError = true;
-      }
-    });
-
-    if (
-      !formData.classTitle ||
-      !formData.description ||
-      !formData.price ||
-      !formData.duration ||
-      !formData.ages.minAge ||
-      !formData.ages.maxAge ||
-      !formData.type ||
-      !formData.lessonType ||
-      availableTimeError
-    ) {
-      setFormErrors({
-        ...formErrors,
-        classTitle: !formData.classTitle
-          ? 'Please provide your class title'
-          : '',
-        category: !formData.category
-          ? 'Please provide the category of the class'
-          : '',
-        description: !formData.description
-          ? 'Please provide your class description'
-          : '',
-        price: !formData.price ? 'Please provide the class price' : '',
-        duration: !formData.duration ? 'Please provide the duration' : '',
-        minAge: !formData.ages.minAge ? 'Please provide the minimum age' : '',
-        maxAge: !formData.ages.maxAge ? 'Please provide the maximum age' : '',
-        lessonType: !formData.lessonType
-          ? 'Please specify the lesson type'
-          : '',
-        type: !formData.type ? 'Please provide the type of class' : '',
-        availableTime: availableTimeError
-          ? 'Please provide the date and start time for all available times'
-          : '',
-      });
-      return;
-    }
-
-    const postedForm = createMultipartForm(formData);
-
-    try {
-      const response = await updateClassForm(
-        classId,
-        userData.token,
-        postedForm
-      );
-      if (response.status === 200) {
-        setIsLoading(true);
-        const response = await getUserData(userData._id, userData.token);
-        setUserData((userData) => ({
-          ...userData,
-          myClasses: response.data.myClasses,
-        }));
-        returnToClasses();
-        setIsLoading(false);
-        setFormErrors({});
-      }
-    } catch (error) {
-      console.log('error', error.data.error);
-      setFormErrors({ ...formErrors, form: error.data.error });
-    }
-  };
-
-  const createMultipartForm = (data) => {
-    const multipartForm = new FormData();
-
-    for (let key in data) {
-      if (key === 'ages') {
-        multipartForm.append('ages[minAge]', data.ages.minAge);
-        multipartForm.append('ages[maxAge]', data.ages.maxAge);
-      } else if (key === 'availableTime') {
-        data.availableTime.forEach((time, index) => {
-          multipartForm.append(
-            `availableTime[${index}][date]`,
-            new Date(time.date)
-          );
-          multipartForm.append(
-            `availableTime[${index}][startTime]`,
-            time.startTime
-          );
-        });
-      } else if (data[key] !== null && data[key] !== undefined) {
-        multipartForm.append(key, data[key]);
-      }
-    }
-
-    return multipartForm;
-  };
-
   return (
     <>
       {isLoading ? (
@@ -231,7 +73,7 @@ const EditClass = () => {
       ) : (
         <>
           <h1 className="text-black font-semibold text-xl sm:text-2xl font-spartan my-4">
-            Edit class {formData.classTitle}
+            Edit class
           </h1>
           <ClassForm
             onChange={handleChange}
