--- src/util/FormatTimeTo24H.js
@@ -1,11 +1,0 @@
-export const FormatTimeTo24H = (time12h) => {
-  const [time, modifier] = time12h.split(' ');
-  let [hours, minutes] = time.split(':');
-  if (hours === '12') {
-    hours = '00';
-  }
-  if (modifier === 'PM') {
-    hours = parseInt(hours, 10) + 12;
-  }
-  return `${hours}:${minutes}`;
-};
